% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsheet_id_from_name.R
\name{get_gsheet_id_from_name}
\alias{get_gsheet_id_from_name}
\title{Get Google Sheet id by name}
\usage{
get_gsheet_id_from_name(gsheet_name)
}
\arguments{
\item{gsheet_name}{character string of length one of the exact match}
}
\description{
Get Google Sheet id by name
}
