% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsheet_metadata.R
\name{get_gsheet_metadata}
\alias{get_gsheet_metadata}
\title{Gets the Google Sheet metadata}
\usage{
get_gsheet_metadata(gsheet_id)
}
\arguments{
\item{gsheet_id}{id}
}
\description{
Gets the Google Sheet metadata
}
