% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_append.R
\name{gs_append}
\alias{gs_append}
\title{Adds a dataframe to a Google Sheet tab by doing a dplyr::bind_row on the data downloaded as a tmp csv and completely rewriting the appended data to a new tab and deleting the old tab}
\usage{
gs_append(dataframe, gsheet_name, tab)
}
\description{
Adds a dataframe to a Google Sheet tab by doing a dplyr::bind_row on the data downloaded as a tmp csv and completely rewriting the appended data to a new tab and deleting the old tab
}
