% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_colnames.R
\name{gs_colnames}
\alias{gs_colnames}
\title{Get colnames every tab in Google Sheet}
\usage{
gs_colnames(gsheet_name, tab, system.sleep = 2)
}
\arguments{
\item{gsheet_name}{name of Google Sheet}
}
\description{
Get colnames every tab in Google Sheet
}
