% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_download_as_csv.R
\name{gs_download_as_csv}
\alias{gs_download_as_csv}
\title{Downloads all the tabs in a given Google Sheet as csvs to destination path with filename as the tab name}
\usage{
gs_download_as_csv(gsheet_name, dest_path, prefix_csv_fn, system.sleep = 2)
}
\description{
Downloads all the tabs in a given Google Sheet as csvs to destination path with filename as the tab name
}
