% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_id_from_name.R
\name{gs_id_from_name}
\alias{gs_id_from_name}
\title{Get Google Sheet id by name}
\usage{
gs_id_from_name(gsheet_name)
}
\arguments{
\item{gsheet_name}{character string of length one of the exact match}
}
\description{
Get Google Sheet id by name
}
