% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ss_by_name.R
\name{gs_ss_by_name}
\alias{gs_ss_by_name}
\title{Gets the Google Sheet metadata}
\usage{
gs_ss_by_name(gsheet_name)
}
\arguments{
\item{gsheet_name}{name of the gsheet as an exact string match. An error will be thrown if there are more than 1 Google Sheets with that name.}
}
\description{
Gets the Google Sheet metadata
}
