% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ss_by_names.R
\name{gs_ss_by_names}
\alias{gs_ss_by_names}
\title{Gets a list of googlesheet metadata from a list of gsheet names (exact string match)}
\usage{
gs_ss_by_names(gsheet_names)
}
\arguments{
\item{gsheet_names}{character vector of gsheet names that are an exact string match}
}
\description{
Gets a list of googlesheet metadata from a list of gsheet names (exact string match)
}
