% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_tab_names.R
\name{gs_tab_names}
\alias{gs_tab_names}
\title{Get all the tab names in the Google Sheet}
\usage{
gs_tab_names(gsheet_name, system.sleep = 0)
}
\arguments{
\item{gsheet_name}{exact Google Sheet name}
}
\description{
Get all the tab names in the Google Sheet
}
