% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_tabnames.R
\name{gs_tabnames}
\alias{gs_tabnames}
\title{Get all the tab names in the Google Sheet}
\usage{
gs_tabnames(gsheet_metadata)
}
\arguments{
\item{gsheet_metadata}{gsheet_metadata}
}
\description{
Get all the tab names in the Google Sheet
}
