% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_write_with_temp_csv.R
\name{gs_write_with_temp_csv}
\alias{gs_write_with_temp_csv}
\title{Write to new Google Sheet}
\usage{
gs_write_with_temp_csv(data, new_gs_name, system.sleep = 0.3)
}
\arguments{
\item{new_gs_name}{name of new gs}

\item{system.sleep}{time in seconds for system sleep}

\item{csv_filename}{full path to file to upload}
}
\description{
Write to new Google Sheet
}
