% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunit-package.R
\docType{package}
\name{gunit}
\alias{gunit}
\alias{gunit-package}
\title{\code{gunit} package}
\description{
Convert Conductance Units
}
\details{
See the README on
\href{https://github.com/cdmuir/gunit}{GitHub}
}
