% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-species.R
\name{gw2gc}
\alias{gw2gc}
\alias{gc2gw}
\title{Convert g_c (\eqn{\mu}mol CO2/m^2/s/Pa) to g_w (\eqn{\mu}mol H2O /m^2/s/Pa)}
\usage{
gw2gc(g_w, D_c, D_w, unitless, a)

gc2gw(g_c, D_c, D_w, unitless, a)
}
\arguments{
\item{g_w}{conductance to water vapor in units
(\eqn{\mu}mol H2O / (m^2 s Pa)) of class \code{units}.}

\item{D_c}{diffusion coefficient for CO2 in air in units of m^2/s of class
\code{units}}

\item{D_w}{diffusion coefficient for H2O in air in units of m^2/s of class
\code{units}}

\item{unitless}{Logical. Should scientific units of arguments be checked and set? TRUE is safer, but slower. If FALSE, values provided are assumed to be in correct units.

\code{units}}

\item{a}{exponent used for conversion. Use 1 for still air; 0.67 for laminar flow (Jones 2014). Should be unitless.}

\item{g_c}{conductance to CO2 in units (\eqn{\mu}mol H2O / (m^2 s Pa)) of
class \code{units}.}
}
\value{
Value with units \eqn{\mu}mol / (m^2 s Pa) of class \code{units}.
}
\description{
Convert g_c (\eqn{\mu}mol CO2/m^2/s/Pa) to g_w (\eqn{\mu}mol H2O /m^2/s/Pa)

Convert g_c (umol CO2/m^2/s/Pa) to g_w (umol H2O /m^2/s/Pa)
}
\details{
Diffusive conductance to CO2 is greater than that of H2O because of the
higher molecular weight. To convert:

\deqn{g_\mathrm{c} = g_\mathrm{w} (D_\mathrm{c} / D_\mathrm{w}) ^ a}{g_c = g_w (D_c / D_w) ^ a}
\deqn{g_\mathrm{w} = g_\mathrm{c} (D_\mathrm{w} / D_\mathrm{c}) ^ a}{g_w = g_c (D_w / D_c) ^ a}
}
\note{
This function will soon be moving to the standalone gunit package.
}
\examples{
library(units)
D_c = set_units(1.29e-05, "m^2/s")
D_w = set_units(2.12e-05, "m^2/s")
g_c = set_units(3, "umol/m^2/s/Pa")
a = 1
g_w = gc2gw(g_c, D_c, D_w, a, unitless = FALSE)
g_w

gw2gc(g_w, D_c, D_w, a, unitless = FALSE)
}
\references{
Jones H. 2014. Plants and Microclimate (3rd edition). Cambridge University Press.
}
