% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol_gs.R
\name{ncol_gs}
\alias{ncol_gs}
\title{Get number of cols in each tab}
\usage{
ncol_gs(gsheet_metadata)
}
\arguments{
\item{gsheet_metadata}{object derived from the}
}
\description{
Get number of cols in each tab
}
