% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gs.R
\name{read_gs}
\alias{read_gs}
\title{Reads Google Sheet as a list of dataframes or a single dataframe}
\usage{
read_gs(gsheet_metadata, system.sleep = 2)
}
\arguments{
\item{gsheet_metadata}{gsheet_metadata}
}
\description{
Reads Google Sheet as a list of dataframes or a single dataframe
}
