% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gs_by_tmp_csv.R
\name{read_gs_by_tmp_csv}
\alias{read_gs_by_tmp_csv}
\title{Reads Google Sheet as a list of dataframes or a single dataframe}
\usage{
read_gs_by_tmp_csv(gsheet_name, tab = NULL, system.sleep = 3)
}
\arguments{
\item{gsheet_name}{gsheet_name}
}
\description{
Reads Google Sheet as a list of dataframes or a single dataframe
}
