% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gs_with_id_refresh.R
\name{read_gs_with_id_refresh}
\alias{read_gs_with_id_refresh}
\title{Reads a googlesheet whilst also replacing NA or blank "TIMESTAMP" values as current timestamp}
\usage{
read_gs_with_id_refresh(gsheet_id)
}
\arguments{
\item{gsheet_id}{google sheet id}
}
\description{
Reads a googlesheet whilst also replacing NA or blank "TIMESTAMP" values as current timestamp
}
