% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_data_gs.R
\name{refresh_data_gs}
\alias{refresh_data_gs}
\title{Refreshes the object with the newest data in the gsheet}
\usage{
refresh_data_gs(gsheet_id, robj_name)
}
\arguments{
\item{gsheet_id}{gsheet_id}

\item{robj_name}{name of object whose data needs to be refreshed}
}
\description{
Refreshes the object with the newest data in the gsheet
}
