% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_gs_colnames.R
\name{retrieve_gs_colnames}
\alias{retrieve_gs_colnames}
\title{Get colnames every tab in Google Sheet}
\usage{
retrieve_gs_colnames(gsheet_name, system.sleep = 2)
}
\arguments{
\item{gsheet_name}{name of Google Sheet}
}
\description{
Get colnames every tab in Google Sheet
}
