% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_gs_fields_with_str.R
\name{retrieve_gs_fields_with_str}
\alias{retrieve_gs_fields_with_str}
\title{Get structure of every tab in Google Sheet}
\usage{
retrieve_gs_fields_with_str(gsheet_name, system.sleep = 2)
}
\arguments{
\item{gsheet_name}{name of Google Sheet}
}
\description{
Get structure of every tab in Google Sheet
}
