% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_gs_fields_with_summary.R
\name{retrieve_gs_fields_with_summary}
\alias{retrieve_gs_fields_with_summary}
\title{Get summary of every tab in Google Sheet}
\usage{
retrieve_gs_fields_with_summary(gsheet_name, system.sleep = 2)
}
\arguments{
\item{gsheet_name}{name of Google Sheet}
}
\description{
Get summary of every tab in Google Sheet
}
