% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_gsheet_id_by_name.R
\name{retrieve_gsheet_id_by_name}
\alias{retrieve_gsheet_id_by_name}
\title{Get Google Sheet id by name}
\usage{
retrieve_gsheet_id_by_name(gsheet_name)
}
\arguments{
\item{gsheet_name}{character string of length one of the exact match}
}
\description{
Get Google Sheet id by name
}
