% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnames_gs.R
\name{tabnames_gs}
\alias{tabnames_gs}
\title{Get all the tab names in the Google Sheet}
\usage{
tabnames_gs(gsheet_metadata)
}
\arguments{
\item{gsheet_metadata}{gsheet_metadata}
}
\description{
Get all the tab names in the Google Sheet
}
