% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gs_ss_by_name.R
\name{update_gs_ss_by_name}
\alias{update_gs_ss_by_name}
\title{Updates a the list of Google Sheets metadata by the name of the Google Sheet}
\usage{
update_gs_ss_by_name(gs_ss_list, gsheet_name)
}
\arguments{
\item{gs_ss_list}{list of Google Sheets metadata named by the Google Sheet name}

\item{gsheet_name}{name of the Google Sheet where the ss needs to be updated}
}
\description{
Updates a the list of Google Sheets metadata by the name of the Google Sheet
}
