% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gs.R
\name{write_gs}
\alias{write_gs}
\title{Write to new Google Sheet}
\usage{
write_gs(new_gs_name, data, rm_Sheet1 = TRUE)
}
\arguments{
\item{new_gs_name}{character string of the new Google Sheet name}

\item{data}{data to be written}

\item{rm_Sheet1}{should Sheet1 be deleted?}
}
\description{
Write to new Google Sheet
}
