% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gs_by_row.R
\name{write_gs_by_row}
\alias{write_gs_by_row}
\title{Write to new Google Sheet}
\usage{
write_gs_by_row(new_gs_name, data, rm_Sheet1 = TRUE,
  system.sleep = 0.3)
}
\arguments{
\item{new_gs_name}{character string of the new Google Sheet name}

\item{data}{data to be written}

\item{rm_Sheet1}{should Sheet1 be deleted?}

\item{system.sleep}{time in seconds for system sleep}
}
\description{
Write to new Google Sheet
}
