% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_upload_csv.R
\name{gs_upload_csv}
\alias{gs_upload_csv}
\title{Write to new Google Sheet}
\usage{
gs_upload_csv(csv_filename, new_gs_name, system.sleep = 0.3)
}
\arguments{
\item{csv_filename}{full path to file to upload}

\item{new_gs_name}{name of new gs}

\item{system.sleep}{time in seconds for system sleep}
}
\description{
Write to new Google Sheet
}
